Update NuGet package to version 1.3.0
Updated the `<Version>` property in the project file from `1.2.0` to `1.3.0`. Added `<PackageReadmeFile>` and `<PackageIcon>` properties to include a readme file (`README.md`) and an icon (`168750409.png`) in the NuGet package.

Included additional files in the NuGet package:
- `..\README.md` with `<Pack>` set to `True` and `<PackagePath>` set to the root directory.
- `C:\Users\AKR\Downloads\168750409.png` with `<Pack>` set to `True` and `<PackagePath>` set to the root directory.

These changes enhance the NuGet package metadata and ensure the specified files are packaged correctly.